requirements.txt
uvicorn
langchain
python-box
streamlit
langchain-groq
langchain_community
python-dotenv
google-generativeai